remove subclassing due to missing support in Lombok (see: https://github.com/rzwitserloot/lombok/issues/853 and https://github.com/rzwitserloot/lombok/pull/1337) to ensure fluent usage patterns illustrated in GrafanaClientDashboardIntegrationTest continue to work
