--- src/main/java/com/appnexus/grafana/client/models/AbsoluteDashboardPanelLink.java
@@ -1,15 +1,0 @@
-/* Licensed under Apache-2.0 */
-package com.appnexus.grafana.client.models;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-import lombok.experimental.Accessors;
-
-@Data
-@EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
-@Accessors(fluent = true)
-public class AbsoluteDashboardPanelLink extends DashboardPanelLink {
-  private String url;
-}
