+++ src/main/java/com/appnexus/grafana/client/models/DashboardAnnotationList.java
@@ -1,7 +1,6 @@
 /* Licensed under Apache-2.0 */
 package com.appnexus.grafana.client.models;
 
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -13,9 +12,6 @@ import java.util.List;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @Accessors(fluent = true)
-@JsonDeserialize(using = DashboardAnnotationListDeserializer.class)
 public class DashboardAnnotationList extends FlexibleSchemaComponent {
-  final static String LIST_FIELD_NAME = "list";
-
   private List<DashboardAnnotation> list;
 }
