+++ src/main/java/com/appnexus/grafana/client/models/DashboardPanelLink.java
@@ -1,9 +1,6 @@
 /* Licensed under Apache-2.0 */
 package com.appnexus.grafana.client.models;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonSubTypes.Type;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -14,20 +11,17 @@ import lombok.experimental.Accessors;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @Accessors(fluent = true)
-@JsonTypeInfo(
-    use = JsonTypeInfo.Id.NAME,
-    include = JsonTypeInfo.As.PROPERTY,
-    property = "type")
-@JsonSubTypes({
-    @Type(value = AbsoluteDashboardPanelLink.class, name = "absolute"),
-    @Type(value = RelativeDashboardPanelLink.class, name = "dashboard")
-})
-abstract public class DashboardPanelLink extends FlexibleSchemaComponent {
+public class DashboardPanelLink extends FlexibleSchemaComponent {
+  // common
   private Type type;
   private Boolean keepTime;
   private String params;
   private Boolean targetBlank;
   private String title;
+  // DASHBOARD
+  private String dashboard;
+  // ABSOLUTE
+  private String url;
 
   public enum Type {
     DASHBOARD("dashboard"),
