+++ src/main/java/com/appnexus/grafana/client/models/Datasource.java
@@ -1,0 +1,5 @@
+package com.appnexus.grafana.client.models;
+
+final public class Datasource {
+  public static final String BUILTIN_DATASOURCE_NAME = "-- Grafana --";
+}
