--- src/main/java/com/appnexus/grafana/client/models/GraphiteDashboardAnnotation.java
@@ -1,18 +1,0 @@
-/* Licensed under Apache-2.0 */
-package com.appnexus.grafana.client.models;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-import lombok.experimental.Accessors;
-
-import java.util.List;
-
-@Data
-@EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
-@Accessors(fluent = true)
-public class GraphiteDashboardAnnotation extends DashboardAnnotation {
-  private List<String> tags;
-  private String target;
-}
